Rewrite the classes and add the aggregating locator
