+++ src/Resources/config/services.yml
@@ -43,31 +43,35 @@ services:
             # Priority must be lower than the one of the initialize_system listener.
             - { name: kernel.event_listener, event: kernel.request, method: onKernelRequest, priority: 26 }
 
-    contao.resource_locator.bundles:
-        class: Contao\CoreBundle\Config\FileLocator
+    contao.resource_locator:
+        alias: contao.resource_locator.chain
+
+    contao.resource_locator.aggregate:
+        class: Contao\CoreBundle\Config\FileLocatorAggregate
         public: false
-        factory: [Contao\CoreBundle\Config\FileLocator, createFromKernelBundles]
         arguments:
-            - "@kernel"
+            - ["@contao.resource_locator.bundles"]
         tags:
             - { name: contao.resource_locator }
+
+    contao.resource_locator.bundles:
+        class: Contao\CoreBundle\Config\BundleFileLocator
+        public: false
+        factory: [Contao\CoreBundle\Config\BundleFileLocator, createFromKernelBundles]
+        arguments:
+            - "@kernel"
+
+    contao.resource_locator.cache:
+        class: Symfony\Component\Config\FileLocator
+        public: false
+        arguments:
+            - ['%kernel.cache_dir%/contao']
+        tags:
+            - { name: contao.resource_locator, priority: 32 }
 
     contao.resource_locator.chain:
         class: Contao\CoreBundle\Config\ChainFileLocator
         public: false
-
-    contao.resource_locator.cache:
-        class: Contao\CoreBundle\Config\CombinedFileLocator
-        public: false
-        arguments:
-            - '%kernel.cache_dir%/contao'
-            - "@contao.resource_locator"
-
-    contao.resource_locator:
-        alias: contao.resource_locator.bundles
-
-    contao.cached_resource_locator:
-        alias: contao.resource_locator.cache
 
     contao.routing.frontend_loader:
         class: Contao\CoreBundle\Routing\FrontendLoader
